refactor(db): split modules
